The setup screen can now navigate to the game scene or back to the main menu.
